requirements.txt
Flask==3.1.2
WeasyPrint
Pillow
qrcode[pil]
PyJWT==2.12.1
Flask-Cors==5.0.0
psycopg2-binary==2.9.11
python-dotenv==1.2.1
Werkzeug==3.1.4
Jinja2==3.1.6
itsdangerous==2.2.0
click==8.1.8
blinker==1.9.0
# Production additions
gunicorn==23.0.0
Flask-Limiter==3.9.0
boto3==1.38.0